code cleanup + comments, risk class moved to new file
